phonebook/requirements.txt
black
Django
django-phonenumber-field[phonenumberslite]
djangorestframework
djangorestframework-simplejwt
drf-yasg
isort
waitress